backend/pyproject.toml
[project]
name = "nexuspkm"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "apscheduler>=3.10",
    "httpx>=0.27.0",
    "llama-index-core>=0.12",
    "llama-index-llms-bedrock>=0.4",
    "llama-index-llms-openai>=0.4",
    "llama-index-llms-ollama>=0.4",
    "llama-index-embeddings-bedrock>=0.4",
    "llama-index-embeddings-openai>=0.4",
    "llama-index-embeddings-ollama>=0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.6.0",
    "mypy>=1.11",
    "types-pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nexuspkm"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["I", "E", "W", "F", "UP", "B", "SIM", "PT"]

[tool.mypy]
python_version = "3.12"
strict = true

[[tool.mypy.overrides]]
module = "llama_index.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "llama_index.*.*"
ignore_missing_imports = true
